feat: Add integration tests for authentication flows
This change introduces a suite of Playwright integration tests to cover various user authentication and enrollment flows.

Key changes include:
- A new test for the standard web client enrollment flow, including TOTP setup.
- New tests for successful and failed UAPI-based authentication flows.
- A new, working test for a native application authentication flow using PKCE.
- Consolidation of all tests into a single file (`test/all_tests.spec.mjs`) to improve test execution stability.
- Addition of a mock authentication layer in `server/oidc/account.js` to facilitate testing without requiring a live UAPI connection. This mock is only active when `NODE_ENV=test`.
- Addition of a `/health` endpoint to the client server and configuration of Playwright to use it, ensuring the server is fully ready before tests run.
